backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
google-generativeai
python-multipart
passlib[bcrypt]
bcrypt==3.2.0
python-jose[cryptography]
google-auth
google-auth-oauthlib
google-auth-httplib2
mercadopago